used the same screen for sign in and sign up
